criterion benchmarking removed, benchmarking using std::time
