--- hash/risc0/bench/.gitignore
@@ -1,4 +1,0 @@
-.DS_Store
-Cargo.lock
-methods/guest/Cargo.lock
-target/
