--- hash/risc0/bench/.vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "rust-analyzer.linkedProjects": [
-        "./methods/guest/Cargo.toml"
-    ]
-}
