+++ hash/risc0/bench/host/Cargo.toml
@@ -7,12 +7,7 @@ edition = "2021"
 methods = { path = "../methods" }
 risc0-zkvm = { version = "0.19.0" }
 serde = "1.0"
-rand = "0.8.5"
-clap = {version = "4.4.8", features = [ "derive" ]}
+rand = "0.8.3"
 
 [dev-dependencies]
-criterion = {version ="0.5.1", default-features = false}
-
-[[bench]]
-name = "bench_main"
-harness = false
+criterion = "0.5.1"
