--- hash/risc0/bench/host/benches/benchmarks/mod.rs
@@ -1,1 +1,0 @@
-pub mod sha256_benchmarks;
