+++ hash/risc0/bench/methods/guest/Cargo.toml
@@ -8,3 +8,7 @@ edition = "2021"
 [dependencies]
 # If you want to try (experimental) std support, add `features = [ "std" ]` to risc0-zkvm
 risc0-zkvm = { version = "0.19.0", default-features = false, features = [ "std" ] }
+
+# [lib]
+# name = "guest"
+# path = "src/mod.rs"
